me pondre el uniforme de piel humana

+++ src/main/java/com/example/proyectodbp/bus/domain/BusService.java
@@ -6,15 +6,11 @@ import com.example.proyectodbp.bus.infraestructure.BusRepository;
 import com.example.proyectodbp.events.HelloEmailEvent;
 import com.example.proyectodbp.exceptions.ResourceNotFoundException;
 import com.example.proyectodbp.exceptions.UnauthorizedOperationException;
+import com.example.proyectodbp.exceptions.UniqueResourceAlreadyExist;
 import com.example.proyectodbp.route.domain.Route;
 import com.example.proyectodbp.route.infraestructure.RouteRepository;
-import com.example.proyectodbp.user.domain.Role;
-import com.example.proyectodbp.user.domain.User;
-import com.example.proyectodbp.user.infraestructure.UserRepository;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import com.example.proyectodbp.utils.AuthorizationUtils;
 
@@ -22,43 +18,29 @@ import com.example.proyectodbp.utils.AuthorizationUtils;
 public class BusService {
     private final BusRepository busRepository;
     private final RouteRepository routeRepository;
-    private final UserRepository<User> userRepository;
     private final AuthorizationUtils authorizationUtils;
-    private final ModelMapper modelMapper;
+    private final ModelMapper modelMapper = new ModelMapper();
 
     private final ApplicationEventPublisher applicationEventPublisher;
 
-    @Autowired
-    public BusService(BusRepository busRepository, RouteRepository routeRepository, UserRepository<User> userRepository, AuthorizationUtils authorizationUtils, ApplicationEventPublisher applicationEventPublisher) {
+    public BusService(BusRepository busRepository, RouteRepository routeRepository, AuthorizationUtils authorizationUtils, ApplicationEventPublisher applicationEventPublisher) {
         this.busRepository = busRepository;
         this.routeRepository = routeRepository;
-        this.userRepository = userRepository;
         this.authorizationUtils = authorizationUtils;
         this.applicationEventPublisher = applicationEventPublisher;
-        this.modelMapper = new ModelMapper();
     }
   
     public String createBus(NewBusRequestDto busDto) {
         if (busRepository.findByPlate(busDto.getPlate()).isPresent()) {
-            throw new ResourceNotFoundException("This bus already exists");
+            throw new UniqueResourceAlreadyExist("This bus already exists");
         }
         Bus newBus = modelMapper.map(busDto, Bus.class);
         return "/driver/"+newBus.getId();
     }
 
     public BusDto getBus(Long id) {
-        // Verifica si el usuario actual es un administrador o el propietario del recurso
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
-
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
         Bus bus = busRepository
                 .findById(id)
@@ -68,22 +50,15 @@ public class BusService {
     }
 
     public void deleteBus(Long id) {
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
+
         busRepository.deleteById(id);
     }
 
     public void updateBus(Long id, BusDto busDto) {
-        // Check if the current user is an admin or the owner of the resource
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
-
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
         Bus busToUpdate = busRepository
                 .findById(id)
@@ -96,18 +71,8 @@ public class BusService {
     }
 
     public void updateBusRoute(Long id, String routeName) {
-        // Check if the current user is an admin or the owner of the resource
-        String username = authorizationUtils.getCurrentUserEmail();
-        if(username == null) {
-            throw new UnauthorizedOperationException("Anonymous User not allowed to access");
-        }
-
-        // Verifica que el usuario actual sea un DRIVER
-        User user = userRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        if(user.getRole() != Role.DRIVER) {
-            throw new UnauthorizedOperationException("No estas autorizado para acceder a este recurso");
-        }
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizedOperationException("User has no permission to modify this resource");
 
         Bus bus = busRepository
                 .findById(id)
